Fixes bug where Vary header was omitted in certain occasions
As explained in
http://www.fastly.com/blog/best-practices-for-using-the-vary-header/
responses should ALWAYS include `Vary: Origin` if there's a chance for
the response to differ depending on the Origin header in the request.

This covers the case for when 2 requests for the same resource, one
including the Origin header and the other not including it, should have
different responses. If the Vary header doesn't mention Origin, then
intermediary caches (like any CDN) will cache the response (including
its headers) and will use it regardless of the request including the
Origin header or not.

This is to play nice with caches. Otherwise rack-cors can't be used with
caches in certain situations.